refactor : delete SeatService & Add to MusicalServcie
