+++ src/main/java/com/bootcamp/savemypodo/service/MusicalService.java
@@ -2,9 +2,12 @@ package com.bootcamp.savemypodo.service;
 
 import com.bootcamp.savemypodo.dto.musical.MusicalResponse;
 import com.bootcamp.savemypodo.entity.Musical;
+import com.bootcamp.savemypodo.entity.Seat;
 import com.bootcamp.savemypodo.global.enums.SortType;
 import com.bootcamp.savemypodo.repository.MusicalRepository;
 import com.bootcamp.savemypodo.repository.ReservationRepository;
+import com.bootcamp.savemypodo.repository.SeatRepository;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +19,7 @@ import java.util.stream.Collectors;
 public class MusicalService {
     private final MusicalRepository musicalRepository;
     private final ReservationRepository reservationRepository;
+    private final SeatRepository seatRepository;
 
     public List<MusicalResponse> getPerformances(SortType sort, Long userId) {
         List<Musical> musicals;
@@ -45,10 +49,13 @@ public class MusicalService {
 			        return MusicalResponse.fromEntity(musical, isReserved); 
 			    })
 			    .collect(Collectors.toList());
-        /*
-        return musicals.stream()
-                .map(MusicalResponse::fromEntity)
-                .collect(Collectors.toList());*/
+    }
+    
+    public List<Seat> getSeatsByMusicalId(Long musicalId) {
+        Musical musical = musicalRepository.findById(musicalId)
+                .orElseThrow(() -> new IllegalArgumentException("공연이 존재하지 않습니다."));
+
+        return seatRepository.findByMusical_Id(musicalId);
     }
 }
